The Point struct should use int instead of ushort so that out of bounds values will be supported.

+++ Voxel2Pixel/Model/Records.cs
@@ -1,6 +1,6 @@
 ﻿namespace Voxel2Pixel.Model
 {
-	public readonly record struct Point(ushort X, ushort Y);
+	public readonly record struct Point(int X, int Y);
 	public readonly record struct Voxel(ushort X, ushort Y, ushort Z, byte Index);
 	public enum VisibleFace : byte
 	{
